pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "future-news-forecaster"
version = "0.1.0"
description = "MVP pipeline for forecasting calendar-driven news headlines and leads"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "openai>=2.14.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "requests>=2.31.0",
]

[project.scripts]
future-news = "future_news_forecaster.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
